+++ spec/hatchet/pipenv_spec.rb
@@ -9,7 +9,7 @@ RSpec.shared_examples 'builds using Pipenv with the requested Python version' do
         remote: -----> Python app detected
         remote: -----> Using Python version specified in Pipfile.lock
         remote: -----> Installing python-#{python_version}
-        remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+        remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
         remote: -----> Installing dependencies with Pipenv 2023.2.4
         remote:        Installing dependencies from Pipfile.lock \\(.+\\)...
         remote: -----> Installing SQLite3
@@ -43,7 +43,7 @@ RSpec.describe 'Pipenv support' do
           remote: -----> No Python version was specified. Using the buildpack default: python-#{DEFAULT_PYTHON_VERSION}
           remote:        To use a different version, see: https://devcenter.heroku.com/articles/python-runtimes
           remote: -----> Installing python-#{DEFAULT_PYTHON_VERSION}
-          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
           remote: -----> Installing dependencies with Pipenv 2023.2.4
           remote:        Installing dependencies from Pipfile...
           remote: -----> Installing SQLite3
@@ -62,7 +62,7 @@ RSpec.describe 'Pipenv support' do
           remote: -----> No Python version was specified. Using the buildpack default: python-#{DEFAULT_PYTHON_VERSION}
           remote:        To use a different version, see: https://devcenter.heroku.com/articles/python-runtimes
           remote: -----> Installing python-#{DEFAULT_PYTHON_VERSION}
-          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
           remote: -----> Installing dependencies with Pipenv 2023.2.4
           remote:        Installing dependencies from Pipfile.lock \\(.+\\)...
           remote: -----> Installing SQLite3
@@ -196,7 +196,7 @@ RSpec.describe 'Pipenv support' do
             remote:  !     See: https://devcenter.heroku.com/articles/python-runtimes
             remote:  !     
             remote: -----> Installing python-#{LATEST_PYTHON_3_7}
-            remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+            remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
             remote: -----> Installing dependencies with Pipenv 2023.2.4
             remote:        Installing dependencies from Pipfile.lock \\(.+\\)...
             remote: -----> Installing SQLite3
@@ -250,7 +250,7 @@ RSpec.describe 'Pipenv support' do
           remote: -----> Python app detected
           remote: -----> Using Python version specified in Pipfile.lock
           remote: -----> Installing python-#{LATEST_PYTHON_3_11}
-          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
           remote: -----> Installing dependencies with Pipenv 2023.2.4
           remote:        Installing dependencies from Pipfile.lock \\(.+\\)...
           remote: -----> Installing SQLite3
@@ -273,7 +273,7 @@ RSpec.describe 'Pipenv support' do
           remote:  !     See: https://devcenter.heroku.com/articles/python-runtimes
           remote:  !     
           remote: -----> Installing python-3.10.7
-          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
           remote: -----> Installing dependencies with Pipenv 2023.2.4
           remote:        Installing dependencies from Pipfile.lock \\(.+\\)...
           remote: -----> Installing SQLite3
@@ -322,7 +322,7 @@ RSpec.describe 'Pipenv support' do
           remote: -----> Python app detected
           remote: -----> Using Python version specified in runtime.txt
           remote: -----> Installing python-#{LATEST_PYTHON_3_10}
-          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
           remote: -----> Installing dependencies with Pipenv 2023.2.4
           remote:        Installing dependencies from Pipfile.lock \\(.+\\)...
           remote: -----> Installing SQLite3
@@ -340,7 +340,7 @@ RSpec.describe 'Pipenv support' do
           remote: -----> Python app detected
           remote: -----> Using Python version specified in Pipfile.lock
           remote: -----> Installing python-#{LATEST_PYTHON_3_10}
-          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
           remote: -----> Installing dependencies with Pipenv 2023.2.4
           remote:        Installing dependencies from Pipfile.lock \\(.+\\)...
           remote: -----> Installing SQLite3
@@ -359,7 +359,7 @@ RSpec.describe 'Pipenv support' do
           remote: -----> No Python version was specified. Using the buildpack default: python-#{DEFAULT_PYTHON_VERSION}
           remote:        To use a different version, see: https://devcenter.heroku.com/articles/python-runtimes
           remote: -----> Installing python-#{DEFAULT_PYTHON_VERSION}
-          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.37.1
+          remote: -----> Installing pip 22.3.1, setuptools 63.4.3 and wheel 0.38.4
           remote: -----> Installing dependencies with Pipenv 2023.2.4
           remote:        Your Pipfile.lock \\(.+\\) is out of date. Expected: \\(.+\\).
           remote:        .+